Provision Global and Region Resources needed to support one or more Marketing Site Stacks.
