+++ frontend/apps/marketing/cicd/1-setup/account-resources.yml.erb
